chore: rename files for capitalizing file names
