+++ libs/analytics/src/core/_Ga.ts
