+++ libs/analytics/src/core/_Meta.ts
