+++ libs/analytics/src/react/AnalyticsProvider.tsx
@@ -1,8 +1,8 @@
 import { createContext, useMemo, useRef, type ReactNode } from 'react'
+import { Ga } from '../core/_Ga'
+import { Meta } from '../core/_Meta'
+import { TikTok } from '../core/_Tiktok'
 import { Analytics, AnalyticsPlatforms } from '../core/Analytics'
-import { Ga } from '../core/Ga'
-import { Meta } from '../core/Meta'
-import { TikTok } from '../core/Tiktok'
 import { useCallbackEvent } from '../hooks/useCallbackEvent'
 import { useEffectOnce } from '../hooks/useEffectOnce'
 import { type BaseAnalyticsEvent } from '../types/events'
